backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pillow==11.3.0
pydantic-settings==2.10.1
python-multipart==0.0.20
boto3>=1.35.0
